Changed a few more uses of long to int, and also corrected int to intptr_t

+++ code/icarus/IcarusInterface.h
@@ -138,8 +138,8 @@ public:
 
 	// Save / Load functions
 
-	virtual int		WriteSaveData( unsigned long chid, void *data, int length ) = 0;
-	virtual int		ReadSaveData( unsigned long chid, void *address, int length, void **addressptr = NULL )  = 0;
+	virtual int		WriteSaveData( unsigned int chid, void *data, int length ) = 0;
+	virtual int		ReadSaveData( unsigned int chid, void *address, int length, void **addressptr = NULL )  = 0;
 	virtual int		LinkGame( int gameID, int icarusID ) = 0;
 	
 	// Access functions
